pyproject.toml
[tool.poetry]
name = "odin"
version = "0.1.0"
description = "run with me"
authors = ["matthew <homekeeper89@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.2"
flask = "^2.0.1"
pydantic = "^1.8.2"
requests = "^2.26.0"
flask_jwt_extended = "^4.3.1"
flask-socketio = "^5.1.1"
haversine = "^2.5.1"
Flask-Migrate = "^3.1.0"
sqlalchemy = "^1.4.25"
flask_sqlalchemy = "^2.5.1"
mysql-connector = "^2.2.9"
python-dotenv = "^0.19.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-picked = "^0.4.6"
pytest-testmon = "^1.2.0"
pytest-watch = "^4.2.0"
black = "21.4b2"
isort = "^5.9.3"
ipdb = "^0.13.9"
flasgger = "^0.9.5"
requests-mock = "^1.9.3"
flask-sqlacodegen = "^1.1.8"
pytest-factoryboy = "^2.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
